+++ mne/tests/test_annotations.py
@@ -635,4 +635,19 @@ def test_event_id_function_using_custom_function():
     assert event_id == expected_event_id
 
 
+@pytest.mark.parametrize('meas_date, out', [
+    pytest.param('toto', 0, id='invalid string'),
+    pytest.param(None, 0, id='None'),
+    pytest.param(42, 42.0, id='Scalar'),
+    pytest.param(3.14, 3.14, id='Float'),
+    pytest.param((3, 140000), 3.14, id='Scalar touple'),
+    pytest.param('2002-12-03 19:01:11.720100', 1038942071.7201,
+                 id='valid iso8601 string'),
+    pytest.param('2002-12-03T19:01:11.720100', 0,
+                 id='invalid iso8601 string')])
+def test_handle_meas_date(meas_date, out):
+    """Test meas date formats."""
+    assert _handle_meas_date(meas_date) == out
+
+
 run_tests_if_main()